idk final i guess

+++ Assets/Scripts/MenuManager.cs
@@ -7,15 +7,18 @@ using UnityEngine.SceneManagement;
 public class MenuManager : MonoBehaviour
 {
     [SerializeField] Animator animator;
-    [SerializeField] GameObject continueMenu, newMenu, dialogBox;
+    
+    [SerializeField] GameObject continueMenu, newMenu, dialogBox, loadScreen;
     // Start is called before the first frame update
     void Start()
     {
+        loadScreen.SetActive(false);
+
         Time.timeScale = 1f;
         dialogBox.SetActive(false);
         continueMenu.SetActive(false);
         newMenu.SetActive(false);
-        string path = Application.persistentDataPath + "/playerData.ss";
+        string path = Path.Combine(Application.persistentDataPath, "playerData.ss");
         if (File.Exists(path))
         {
             //file exists. show with continue
@@ -62,15 +65,31 @@ public class MenuManager : MonoBehaviour
     }
     public IEnumerator deleteRoutine() {
 
+        
         animator.SetTrigger("triggerStart");
         yield return new WaitForSeconds(2);
+        loadScreen.SetActive(true);
+        yield return new WaitForSeconds(3);
+        
         string[] filePaths = Directory.GetFiles(Application.persistentDataPath);
         foreach (string filePath in filePaths)
         {
-            File.Delete(filePath);
+
+            if (filePath == Path.Combine(Application.persistentDataPath, "playerData.ss"))
+            {
+                Debug.Log("To delete: "+filePath);
+                File.Delete(filePath);
+            }
+            else { 
+                Debug.Log("Not to: "+filePath);
+
+            }
+
 
         }
-       
+
+        yield return new WaitForSeconds(1);
+        Debug.Log("Going to next scene...");
         SceneManager.LoadScene(1);
         //go to scene 
     }
@@ -79,12 +98,16 @@ public class MenuManager : MonoBehaviour
 
         animator.SetTrigger("triggerStart");
         yield return new WaitForSeconds(2);
+        loadScreen.SetActive(true);
+        yield return new WaitForSeconds(5);
         //go to scene
         SceneManager.LoadScene(1);
     }
     IEnumerator loadGameRoutine() {
         animator.SetTrigger("triggerStart");
         yield return new WaitForSeconds(2);
+        loadScreen.SetActive(true);
+        yield return new WaitForSeconds(5);
         PlayerPrefs.SetInt("isLoad", 1);
         SceneManager.LoadScene(1);
